requirements-dev.txt
shortuuid
werkzeug
django_pdb
python-dateutil
pytest-django
pytest-cov
factory-boy
requests
mock
pygments
vobject

types-pyOpenSSL
types-PyYAML
types-boto
types-requests
